Added KeyCloakPropertiesConfiguration
