+++ backend/src/main/java/hu/flowacademy/worksheet/service/KeycloakClientService.java
@@ -19,7 +19,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Locale;
 
 @RequiredArgsConstructor
 @Service
